requirements.txt
flask_login
flask_bcrypt
flask_sqlalchemy
flask_wtf
flask_migrate
